+++ .gitignore
@@ -2,4 +2,4 @@
 target/
 test-output/
 *properties
-ch-040.iml
+eLibrary.iml
