+++ debug.log
@@ -1,0 +1,10 @@
+[0614/190335.774:ERROR:scoped_process_suspend.cc(30)] NtSuspendProcess: Попытка доступа к процессу, выполняющему процедуру завершения.  (0xc000010a)
+[0614/190335.793:ERROR:process_info.cc(116)] ReadProcessMemory bool __cdecl crashpad::`anonymous-namespace'::ReadStruct<struct crashpad::process_types::PEB<struct crashpad::process_types::internal::Traits64>>(void *,unsigned __int64,struct crashpad::process_types::PEB<struct crashpad::process_types::internal::Traits64> *): Запрос ReadProcessMemory или WriteProcessMemory был выполнен только частично. (0x12B)
+[0614/190335.793:ERROR:process_info.cc(559)] ReadProcessData failed
+[0614/190335.793:ERROR:process_reader_win.cc(233)] ReadMemory at 0x7feeb202968 of 12 bytes failed: Запрос ReadProcessMemory или WriteProcessMemory был выполнен только частично. (0x12B)
+[0614/190335.793:WARNING:process_snapshot_win.cc(68)] ReadMemory ExceptionInformation failed
+[0614/190335.793:WARNING:crash_report_exception_handler.cc(60)] ProcessSnapshotWin::Initialize failed
+[0614/190938.880:ERROR:process_reader_win.cc(142)] SuspendThread: Отказано в доступе. (0x5)
+[0614/190938.881:ERROR:process_reader_win.cc(142)] SuspendThread: Отказано в доступе. (0x5)
+[0614/190938.881:ERROR:exception_snapshot_win.cc(87)] thread ID 5072 not found in process
+[0614/190938.881:WARNING:crash_report_exception_handler.cc(60)] ProcessSnapshotWin::Initialize failed
