+++ src/main/java/elibrary/tools/LocalizationConfig.java
@@ -8,30 +8,11 @@ import java.util.Properties;
 
 public class LocalizationConfig {
 
-
-    public static String setLanguage() {
-        Properties properties = new Properties();
-        try {
-            Reader reader = new InputStreamReader(BaseTest.class.getClassLoader().getResourceAsStream("app.properties"), "UTF-8");
-            properties.load(reader);
-            if ("UA".equals(properties.getProperty("language"))) {
-                return "UA";
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return "EN";
-    }
-
     public static Properties getPropertiesForLocalization() {
         Properties properties = new Properties();
         try {
             Reader reader;
-            if ("UA".equals(setLanguage())) {
-                reader = new InputStreamReader(BaseTest.class.getClassLoader().getResourceAsStream("ua.messages.properties"), "UTF-8");
-            } else {
-                reader = new InputStreamReader(BaseTest.class.getClassLoader().getResourceAsStream("en.messages.properties"), "UTF-8");
-            }
+            reader = new InputStreamReader(Base.class.getClassLoader().getResourceAsStream("ua.messages.properties"), "UTF-8");
             properties.load(reader);
         } catch (UnsupportedEncodingException e) {
             e.printStackTrace();
