+++ src/main/resources/app.properties
@@ -1,6 +1,5 @@
 #
-browserType = ${browser.name}
+browserType = firefox
 driverType = ${driver.name}
 driverPath = ${driver.path}
-language = UA
 #
