+++ src/main/resources/db.properties
@@ -1,7 +1,5 @@
 JDBC_DRIVER = org.postgresql.Driver
-JDBC_URL = jdbc:postgresql://localhost:5433/hospital
-JDBC_MAIN_URL = jdbc:postgresql://localhost:5433
+JDBC_URL = jdbc:postgresql://localhost:5432/elibrary
+JDBC_MAIN_URL = jdbc:postgresql://localhost:5432
 USER = postgres
-PASSWORD = root
-FULL_DATASET_PATH = src/test/resources/dataset_full.xml
-SMALL_DATASET_PATH = src/test/resources/dataset_small.xml
+PASSWORD = pass4Petr0
