+++ src/main/resources/log4j2.xml
@@ -35,7 +35,7 @@
         <Root level="INFO">
             <AppenderRef ref="info-log"/>
         </Root>
-        <Logger name="hospitalSeeker" level="debug" additivity="false">
+        <Logger name="elibrary" level="debug" additivity="false">
             <appender-ref ref="info-log" level="INFO"/>
             <appender-ref ref="error-log" level="ERROR"/>
             <appender-ref ref="console-log" level="INFO"/>
