+++ src/main/resources/ua.messages.properties
@@ -6,4 +6,3 @@ INCORRECT_EMAIL_FORMAT = Будь ласка, введіть коректну а
 INSECURE_PASSWORD = Будь ласка, введіть не менше 6 символів.
 WEAK_PASSWORD = Введений пароль ненадійний.
 CONFIRM_PASSWORD_ERROR = Будь ласка, введіть таке ж значення ще раз.
-BLOCKED_ACCOUNT = Ваш акаунт заблокований.
